Separate requesting/receiving content
This corrects the timer display for queries. It now shows things like
"took 2384ms" instead of silly things like "took 68ms" because it was
measuring the time Ace took instead of the time the Ajax took.

This also fixes #111 and the problem where JTable query results would
flash up and then fade out.