[tests] Refactor tests for easier reading & maintenance
This commit is meant to make the frontend test code a bit easier to
follow and maintain. The main difference is that this version separates
the acquisition of UI elements from the actions that are performed on
them; as such, chains of auxiliary functions are replaced with simple,
non-nested monadic actions.