Refresh after mounting database (details below) [fix #376]
After mounting the database, we should refresh the file listing for the
user. In this commit, this is done by reloading the page, but it would
be nice to just trigger an ajax request for the data itself and have it
be updated. However, I did not go so far as to do that, since the way to
do it seems kind of gnarly: you update the "salt" (why is it called
that?) in the url, which should trigger a new request---but in order to
do this, several more pieces of state are needed which are not
accessible from within `saveMount`. In order to avoid introducing a lot
more plumbing code, I went with a simple refresh for now.

It does not suffice to merely refresh, however, since saving a
mount results in the server restarting itself. In addition to awaiting
the availability of the server (which can be done with a retrying
request to its 'info' endpoint), we very unfortunately have to wait for
the server to go down before we check if it is available! Otherwise, we
may catch the server just before it goes down, and then attempt to
reload the page after the server has gone down.

This is nondeterministic, and it is very dissatisfying to me that
this level of detail about how the backend works is ending up in the
frontend. I have approximated correct behavior by waiting for 1s before
checking if the server is available.

Test notes
==========

I have changed the tests to not start with a preconfigured mount, but
rather to add a mount using the user interface. This is related to #269.
The test that I have added fails without the changes in this pull
request.

Updated the SlamEngine version in the tests to 2.1.6. The old version
that was in here had a bug where it would not restart itself properly. I
am beginning to test creating new mounts, which requires that to work
correctly.

The configuration file may be overwritten in the course of the tests
(e.g. as soon as we test adding a new mount), so I have changed the test
script to backup and restore the slamengine config.