Fix two visual problems (resolves #396)
First, the red alert box in the Mount dialog was not properly resizing
to fit its content.

---

Secondly, in (at least) Safari, the spinners were failing to animate
under some circumstances; for me, it would happen the first time I load
the app after the backend is fired up, and then after a while, they
would animate properly. It is still not clear *why* this was happening,
but at least as far as my own testing is concerned, switching these from
svgs to animated gifs sufficed to resolve the issue.

To generate the animated gifs, I used http://loading.io/ (thanks
@cryogenian); it appears that this is whence the original svgs came in
the first place.