[tests] Test mounting a database with a bad port
It was necessary to upgrade the SlamEngine jar used in the tests to at
least 2.1.7, which changed the mongo timeout to 5s.

In addition, one change in SlamEngine behavior is that it seems to now
be required to specify the name of the database in the connection URI
(mongo://...:12345/testdb). I've updated the test code accordingly.

The added test tries to mount a database with an incorrect port, and
checks that the correct error is presented in the correct way.