Add preventDefault to mount dialog (#400)
This is meant to resolve #400, but sadly I cannot reproduce the
intermittent issue that prompted this change anymore. If indeed it is
the case that the button was causing the page to refresh (by virtue of
being inside a <form>), then this should solve the problem. But it is
not really clear to me that this is what was going on, since I don't
know why it would be non-deterministic.

In any case, the change is probably harmless.

Mount modal form; add preventDefault to onSubmit (via nonSubmit)

Related to #400