Fix several bugs in the tests; upgrade purescript-webdriver
Upgrade purescript-webdriver to version with corrected binding to
`wait`.

There was a subtle bug in Test.Selenium.Monad.checker that caused it to
only loop once, where it was intended to loop forever.

Finally, there were a lot of instances of `waitCheck (later 1000 $ pure
false)...`, which appears to have been intended to introduce a delay.
This actually worked by coincidence with the old broken version of
`wait`, but it fails properly now.

I have replaced these with simple calls to `later`, but we should investigate replacing all of these with proper `waitCheck` conditions that depend on some real aspect of the page. Otherwise, this stuff is just totally non-deterministic and we are merely reducing the frequency of spurious test failure.