[PR Feedback]: Don't test for a particular server error message
One wrinkle is that at the moment we want to wait for the warning box to
appear, the warning box actually is momentarily visible (displaying an
old validation error that will shortly disappear). So it is necessary to
wait for it to be invisible, and then wait for it to become visible
again.