Fix incorrect field pruning logic (fix #420)
The logic as it was would consider *every* field which did not have a
default value as needing to be pruned, since it was comparing against
the result of initSlamDownState, which just creates a dictionary
containing the fields with default values. Instead, it should have been
comparing against the result of slamDownFields.

Thanks to @garyb for muddling through the diagnostic process with me.