[build] Test improvements (resolve SD-446)
Fixed an issue where the wrong keyboard shortcuts were being sent
in the tests. We check the platform in order to decide on key sequences;
in case we are running remote, we want the platform that the client is
running on, not that of the build server.

SD-446 #resolve

Preliminary changes to support remote tests
----------------------------------------------------------------------

- Our test script is not compatible with Node.JS 0.10; I changed the
  version in .travis.yml to 0.12.4.

- Add remote-test gulp task

If we want to run the tests from Travis, we'll need to install a newer
version of MongoDB on Travis. I'll make this change in whatever pull
request enables tests on Travis.

-----------------------------------------------------------------

Also, moved Test.Env into the correct directory.