Add gulp task to automatically trim trailing whitespace
This is just an idea, so if anyone doesn't like it, feel free to close
this.

I've just been noticing that a lot of our source code has trailing
whitespace in it---I have my editor set up to remove it automatically,
but when I do so, it seems to get re-introduced with some regularity.
Not to mention, frequent whitespace-adjustment passes can lead to
difficult to read diffs (and difficult merges), whereas if the invariant
is maintained automatically over time, this won't be a problem.

Adding the removal of trailing whitespace to the build would kind of
solve this for everyone, and nobody would have to make any adjustment to
their editor or be especially careful when preparing a commit.