[travis] Switch to java build w/ Ubuntu Trust Targ environment
Switching to the Java build because this seems to be the easiest way to
get JDK 8 installed, in order to unblock #492. If someone can figure out
how to install it by hand w/ curl, I'm happy to use that approach as
well, but I just have no idea how.

The switch to the "trusty" environment is to be able to easily install
nodejs and npm with apt-get. Additionally, apparently this environment
has 7.5 Gb RAM instead of 3Gb, so we may be able to use this to unblock