SD-1645: Recursive Decks (#765)
This uses the opaque component machinery to break the cycle in types
when trying to make a recursive components. Consequently, Decks can only
be peeked at the top-level, as child queries are now hidden.

To allow child cards to introduce a recursive deck, we thread the
deck's component spec through as an argument.