SD-1614: fix purging of view mounts (#777)
* cleanViewMounts: account for nested directory structure

This resolve SD-1614.

It is not enough to delete the view mounts that occur directly under a
directory; we now have things like
my-workspace.workspace/0/my-view-mount. To resolve this, traverse the
transitive children of the directory looking for view mounts.

This commit also golfs a bit of the code in SlamData.Quasar.FS, using
ExceptT internally in places where it simplifies things. We still of
course expose the simple APIs in Either.

* prism golf